Exercise 5: machine with jumps and comparisons.
The encoding of Op-Codes separates control from arithmetic
instructions, for systematic treatment.

As before, I separate executing one instruction from
executing the sequence of instructions.

At some point I may want to use a state monad in the interpreter,
but for now this is working fine and understanding is ok.