IntCode interpreter : let us have some monads.
It was not difficult to guess that the interpreter could
use some State monad in it. The result takes more lines of code,
but they are shorted lines overall.

Perhaps some lenses would help with the changes that only refer
to the memory, program counter, input, or output. But Not today.